Added Date Format Tester Panel.
