+++ client-app/src/desktop/tabs/formats/FormatsTab.scss
@@ -17,7 +17,7 @@
       }
     }
     td.inputColumn {
-      width: 40%;
+      width: 291px;
     }
     td.indexColumn {
       width: 24px;
